requirements.txt
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.3
loguru==0.7.2
python-dateutil==2.8.2
langgraph==0.0.26
langchain==0.1.9
langchain-openai==0.0.8
fastapi==0.110.0
uvicorn==0.27.1
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.27.0
aiofiles==23.2.1
python-multipart==0.0.9
python-instagram>=1.3.2
SQLAlchemy>=2.0.0
python-jose[cryptography]>=3.3.0
groq>=0.3.0  # We'll use Groq as it has a free tier 